+++ src/keybinds/config.rs
@@ -8,6 +8,8 @@ pub struct KeybindsConfig {
     pub scroll_down_half: Option<Keybind>,
     pub scroll_up_half: Option<Keybind>,
 
+    pub toggle_layout: Option<Keybind>,
+
     pub log_tab: Option<LogTabKeybindsConfig>,
     pub message_popup: Option<MessagePopupKeybindsConfig>,
 }
@@ -71,6 +73,4 @@ pub struct LogTabKeybindsConfig {
     pub push_all_new: Option<Keybind>,
     pub fetch: Option<Keybind>,
     pub fetch_all: Option<Keybind>,
-
-    pub open_help: Option<Keybind>,
 }