+++ src/ui/mod.rs
@@ -39,6 +39,7 @@ pub enum ComponentAction {
     SetPopup(Option<Box<dyn Component>>),
     Multiple(Vec<ComponentAction>),
     RefreshTab(),
+    ToggleLayout,
 }
 
 pub trait Component {
@@ -54,6 +55,14 @@ pub trait Component {
     fn draw(&mut self, f: &mut Frame<'_>, area: Rect) -> Result<()>;
 
     fn input(&mut self, event: Event) -> Result<ComponentInputResult>;
+
+    fn make_main_panel_help(&self) -> Vec<(String, String)> {
+        vec![]
+    }
+
+    fn make_details_panel_help(&self) -> Vec<(String, String)> {
+        vec![]
+    }
 }
 
 #[instrument(level = "trace", name = "draw", skip(f, app))]